Conflicts:
	src/vahdin/VahdinUI.java

+++ src/vahdin/data/Mark.java
@@ -67,8 +67,8 @@ public class Mark implements Item {
         return getId();
     }
 
-    public float getVoteCount() {
-        return this.voteCount; // TODO:
+    public int getVoteCount() {
+        return (int) this.voteCount; // TODO:
     }
 
     public int getId() {
